WIP - forces

+++ src/gpu/gpu_solver.cuh
@@ -1,15 +1,15 @@
 #pragma once
-
-#include "../model.cuh"
+#include "../model.h"
 #include <unordered_map>
+
 namespace GPUSolver {
     class Solver : public World::Solver {
     public:
         Solver() {};
 
 
-        void Solve(float dt,
-            float time,
+        void Solve(double dt,
+            double time,
             std::vector<World::Object>& objects,
             std::vector<World::Force>& forces,
             size_t stepsNumber,
